+++ library/src/validations/custom/customAsync.ts
@@ -21,7 +21,7 @@ export type CustomValidationAsync<TInput> = BaseValidationAsync<TInput> & {
  * @param requirement The async validation function.
  * @param message The error message.
  *
- * @returns A async validation function.
+ * @returns A async validation action.
  */
 export function customAsync<TInput>(
   requirement: (input: TInput) => Promise<boolean>,